Legal-Saathi-main/Legal-Saathi-main/requirements.txt
streamlit
transformers
torch
sentencepiece
accelerate
PyMuPDF
sentence-transformers
openai-whisper
gTTS
pytesseract
opencv-python-headless
Pillow
numpy